create debug api file
