+++ seele/api.go
@@ -8,7 +8,6 @@ package seele
 import (
 	"math/big"
 
-	"github.com/davecgh/go-spew/spew"
 	"github.com/seeleteam/go-seele/common"
 	"github.com/seeleteam/go-seele/common/hexutil"
 	"github.com/seeleteam/go-seele/core"
@@ -140,68 +139,6 @@ func (api *PublicSeeleAPI) GetBlockByHash(request *GetBlockByHashRequest, result
 	return nil
 }
 
-// PublicDebugAPI provides an API to access full node-related information for debug.
-type PublicDebugAPI struct {
-	s *SeeleService
-}
-
-// NewPublicDebugAPI creates a new NewPublicDebugAPI object for rpc service.
-func NewPublicDebugAPI(s *SeeleService) *PublicDebugAPI {
-	return &PublicDebugAPI{s}
-}
-
-// GetBlockRlp retrieves the RLP encoded for of a single block, when height is -1 the chain head is returned
-func (api *PublicDebugAPI) GetBlockRlp(height *int64, result *string) error {
-	block, err := getBlock(api.s.chain, *height)
-	if err != nil {
-		return err
-	}
-
-	blockRlp, err := common.Serialize(block)
-	if err != nil {
-		return err
-	}
-
-	*result = hexutil.BytesToHex(blockRlp)
-	return nil
-}
-
-// PrintBlock retrieves a block and returns its pretty printed form, when height is -1 the chain head is returned
-func (api *PublicDebugAPI) PrintBlock(height *int64, result *string) error {
-	block, err := getBlock(api.s.chain, *height)
-	if err != nil {
-		return err
-	}
-
-	*result = spew.Sdump(block)
-	return nil
-}
-
-// GetTxPoolContent returns the transactions contained within the transaction pool
-func (api *PublicDebugAPI) GetTxPoolContent(input interface{}, result *map[string][]map[string]interface{}) error {
-	txPool := api.s.TxPool()
-	data := txPool.GetProcessableTransactions()
-
-	content := make(map[string][]map[string]interface{})
-	for adress, txs := range data {
-		trans := make([]map[string]interface{}, len(txs))
-		for i, tran := range txs {
-			trans[i] = rpcOutputTx(tran)
-		}
-		content[adress.ToHex()] = trans
-	}
-	*result = content
-
-	return nil
-}
-
-// GetTxPoolTxCount returns the number of transaction in the pool
-func (api *PublicDebugAPI) GetTxPoolTxCount(input interface{}, result *uint64) error {
-	txPool := api.s.TxPool()
-	*result = uint64(txPool.GetProcessableTransactionsCount())
-	return nil
-}
-
 // PublicNetworkAPI provides an API to access network information.
 type PublicNetworkAPI struct {
 	p2pServer      *p2p.Server
@@ -258,8 +195,6 @@ func (api *PublicMinerAPI) Stop(input *string, result *string) error {
 
 	return nil
 }
-
-// private methods
 
 // rpcOutputBlock converts the given block to the RPC output which depends on fullTx
 func rpcOutputBlock(b *types.Block, fullTx bool) (map[string]interface{}, error) {
